pruebas_api/backend/requirements.txt
# pruebas_api/backend/requirements.txt

# Archivo de requisitos para la aplicación backend de Flask
Flask==3.0.3
psycopg2-binary==2.9.9
requests==2.32.3
python-dotenv==1.0.0
# Fin de requirements.txt